update vpc sdk to 2016-04-28
