+++ package.json
@@ -10,7 +10,7 @@
     "clean": "pnpm run -r clean",
     "build": "pnpm run -r build",
     "test": "jest --testTimeout 10000",
-    "test:pop": "jest --testTimeout 10000 packages/srm-aliyun-pop-core --watchAll",
+    "test:pop": "jest --testTimeout 10000 packages/srm-aliyun-pop-core/**/index.test.ts --watchAll",
     "lint": "eslint --ext js,ts,tsx ."
   },
   "devDependencies": {
