+++ packages/srm-aliyun-pop-core/__tests__/index.test.ts
@@ -1,4 +1,5 @@
-import Pop, { RPCClient, ROAClient } from '../src';
+import Pop, { ROAClient } from '../src';
+import _ from 'lodash';
 import PopClient from '@alicloud/pop-core';
 
 require('dotenv').config({ path: require('path').join(__dirname, '.env') });
@@ -21,11 +22,18 @@ test('ROAClient', async () => {
   })()).resolves.toBeUndefined();
 });
 
-test.only('getInitVpcConfigAsFc', async () => {
+test.only('getInitNasConfigAsFc', async () => {
+  const region = 'cn-chengdu';
   const config = {
     accessKeyId: process.env.ACCESS_KEY_ID || '',
     accessKeySecret: process.env.ACCESS_KEY_SECRET || '',
-    endpoint: 'https://devops.cn-hangzhou.aliyuncs.com',
-    apiVersion: '2021-06-25',
+    endpoint: `http://nas.${region}.aliyuncs.com`,
+    apiVersion: '2017-06-26',
   }
+  const codeupClient = new Pop(config);
+  expect(codeupClient instanceof PopClient).toBeTruthy();
+  await codeupClient.getInitNasConfigAsFc({
+    region,
+    rule: 'test-srm-client'
+  });
 })
