+++ packages/srm-aliyun-pop-core/src/vpc-2016-04-28/type.ts
@@ -58,3 +58,8 @@ export interface ICreateVSwitch {
 
 export type ICreateVSwitchResponse = string;
 
+
+export interface IDescribeVSwitchAttributes {
+  region: string;
+  vswitchId: string;
+}
